% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineCorrect.R
\name{baselineCorrect}
\alias{baselineCorrect}
\title{Baseline correction on a TAP object}
\usage{
baselineCorrect(TAPexperiment, gasName)
}
\arguments{
\item{TAPexperiment}{A collection of TAP objects.}

\item{gasName}{The gas in which to apply the function on.}
}
\value{
The TAPexperiment with applied function.
}
\description{
Correct the baseline of a pulse towards zero
}
\details{
The TAP pulse mass spec results may sometimes not reach zero appropriately when no product is produced or when all the reactant is consumed.
In this case, it is nessecary to perform a baseline correction of the respective pulse.  This is slightly different from typical baseline correction because it is only necessary
to shift the pulse vertically such that the baseline portion is at zero.  This method requires a user to know at what time segment the baseline is not changing, i.e., the slope of the pulse
remains constant or flat.  Once baseline has been determined, this function will shift the whole pulse vertically by subtracting a choice of either the min or mean of the baseline time segment.
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")

# plot a baseline corrected pulse
plot(TAPexperiment$AMU40$matrix[[5]])
}
