% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTAPScript.R
\name{buildTAPScript}
\alias{buildTAPScript}
\title{Execute Series of functions}
\usage{
buildTAPScript(TAPexperiment)
}
\arguments{
\item{TAPexperiment}{Data to be applied}
}
\value{
A collection of TAP objects that describe the experiment performed in the reactor.
}
\description{
Execute Series of functions
}
\examples{


}
