% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructFlux.R
\name{constructFlux}
\alias{constructFlux}
\title{TAP Flux Constructor}
\usage{
constructFlux(none = NULL)
}
\arguments{
\item{none}{No value necessary}
}
\value{
Prefilled Flux parameters
}
\description{
Generates prefilled Flux parameters for options.  Does not include prefilled flux
}
\examples{
TAPstruct = constructFlux()
names(TAPstruct)
}
