% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameToMoments.R
\name{dataFrameToMoments}
\alias{dataFrameToMoments}
\title{Data Frame to Moments}
\usage{
dataFrameToMoments(df, momentsObj)
}
\arguments{
\item{df}{An array.}

\item{momentsObj}{An index of values.}
}
\value{
The TAPexperiment with applied function.
}
\description{
Helper function for subsetting data
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")

# plot a baseline corrected pulse
plot(TAPexperiment$AMU40$matrix[, 5])
}
