% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendFlux.R
\name{extendFlux}
\alias{extendFlux}
\title{Pulse Deconvolution}
\usage{
extendFlux(TAPexperiment, gasName, maxTime)
}
\arguments{
\item{TAPexperiment}{A set of collected gas species.}

\item{gasName}{Name of the Gas to be processed.}
}
\value{
Blah
}
\description{
Pulse Deconvolution
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU28$options$baselineStart = 3.9
TAPexperiment$AMU28$options$baselineStart = 4
TAPexperiment$AMU28$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU28")

TAPexperiment$parameters$experiment$pumpProbeSpacing = 0.1

TAPexperiment = pulseDeconvolution(TAPexperiment, "AMU28")

# plot a baseline corrected pulse
plot(TAPexperiment$AMU28$matrix[[5]])
lines(TAPexperiment$AMU28pre$matrix[[5]], col = "green")
lines(TAPexperiment$AMU28post$matrix[[5]], col = "red")

}
