% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasCon.R
\name{gasCon}
\alias{gasCon}
\title{Gas Concentration}
\usage{
gasCon(TAPexperiment, gasName, method = "y")
}
\arguments{
\item{method}{Blah}

\item{TAPobj}{A set of pre-processing parameters based on the Y-Procedure}
}
\value{
Blah
}
\description{
Optimization of the Y-Procedure Smoothing value
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = gasCon(TAPexperiment, "AMU40")
plot(TAPexperiment$AMU40gasCon$matrix[[5]])

TAPexperiment = gasCon(TAPexperiment, "AMU40", method = "g")
plot(TAPexperiment$AMU40gasCon$matrix[[5]])

}
