% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertGraham.R
\name{inertGraham}
\alias{inertGraham}
\title{Inert Graham Calculation}
\usage{
inertGraham(TAPexperiment, gasName, newMass)
}
\arguments{
\item{newMass}{Transformed mass}

\item{TAPobj}{The inert gas flux}
}
\value{
A matrix or vector of transformed inert flux
}
\description{
Scale inert flux by Graham's Law
}
\details{
This funciton is for manipulation of the inert flux to appear as if it has a different mass.  This allows the inert gas to behave as the reactant gas
without any interaction on the catalyst.  This function scales the magnitude and time by the square root of the ratio of masses.  For example, if the
inert gas was Argon (AMU of 40) and the reactant was Oxygen (AMU 32), the inert will be scaled by \eqn{\sqrt{40/32}}.
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = inertGraham(TAPexperiment, "AMU40", 20)
plot(TAPexperiment[["AMU40"]]$matrix[[5]])
lines(TAPexperiment[["AMU40to20"]]$matrix[[5]])

}
