% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initYproc.R
\name{initYproc}
\alias{initYproc}
\title{Init Y-Procedure}
\usage{
initYproc(TAPexperiment, gasName, method = "y")
}
\arguments{
\item{TAPexperiment}{A set of pre-processing parameters based on the Y-Procedure}

\item{gasName}{Blah}

\item{method}{Blah}
}
\value{
Blah
}
\description{
Optimization of the Y-Procedure Smoothing value
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment$AMU40$options$inert = "AMU40"
TAPexperiment = initYproc(TAPexperiment, "AMU40")
plot(TAPexperiment$AMU40$gasScalar)

}
