% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Moments of an TAP Pulse}
\usage{
moments(TAPexperiment, gasName)
}
\arguments{
\item{TAPobj}{A TAP pulse or series of pulses.}
}
\value{
A vector of moments based on the pulse.
}
\description{
Function for determining the moments of a pulse.
}
\details{
The \eqn{n^{th}} moment is defined by:
\deqn{ \int t^n pulse(t) dt}
}
\examples{

TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = moments(TAPexperiment, "AMU40")
df = momentsToDataFrame(TAPexperiment, "AMU40")
plot(df$M1)

}
