% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentsToDataFrame.R
\name{momentsToDataFrame}
\alias{momentsToDataFrame}
\title{Moments to Data frame}
\usage{
momentsToDataFrame(TAPexperiment, gasName)
}
\arguments{
\item{Array}{An array.}

\item{index}{An index of values.}
}
\value{
The TAPexperiment with applied function.
}
\description{
Helper function for subsetting data
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")

# plot a baseline corrected pulse
plot(TAPexperiment$AMU40$matrix[, 5])
}
