% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimYprocSmoothing.R
\name{optimYprocSmoothing}
\alias{optimYprocSmoothing}
\title{Optimization of the Y-Procedure Smoothing value}
\usage{
optimYprocSmoothing(TAPgasConVars)
}
\arguments{
\item{TAPgasConVars}{A set of pre-processing parameters based on the Y-Procedure}
}
\value{
A vector of moments based on the pulse.
}
\description{
Optimization of the Y-Procedure Smoothing value
}
\details{
blah
TAPgasConVars = list("tempPulse" = tempPulse, "omega" = omega, "timeStep" = timeStep, "gasScalar" = gasScalar, "gamma3" = gamma3, "gasConScalar" = gasConScalar)
}
\examples{
data("pulseData")


}
