% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierTAP.R
\name{outlierTAP}
\alias{outlierTAP}
\title{Outlier Removal}
\usage{
outlierTAP(TAPexperiment, gasName)
}
\arguments{
\item{TAPexperiment}{A TAP object of pulse values.}

\item{gasName}{Value}
}
\value{
A vector of smoothed pulse values.
}
\description{
Partition the pulse with an exponential decay from the mode.  Any points above or below 2 standard deviations are set to the median of the partition.
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = outlierTAP(TAPexperiment, "AMU40")

plot(TAPexperiment$AMU40$matrix[[5]])
}
