% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPulse.R
\name{plotPulseTesting}
\alias{plotPulseTesting}
\title{Plot Pulses}
\usage{
plotPulseTesting(TAPexperiment, xNames, yNames, timeIndex, pulseIndex,
  yIndex = NULL, xNorm = "none", yNorm = "none", xTrans = "none",
  yTrans = "none", newXlab = "", newYlab = "", colRamp = F,
  showLegend = F, plotType = "2D")
}
\arguments{
\item{newXlab}{The x axis label.}

\item{newYlab}{The y axis label.}

\item{colRamp}{A boolean value if the pulses are colored by a single color or matlab like gradient.}

\item{showLegend}{A boolean value to display the legend.}

\item{plotType}{A character describing the plot type.  Options include: "2D", "3D", "Sequence" and "Heatmap"}

\item{xPulses}{A list of data frames containing the pulses to be plotted.}

\item{xTime}{A list of vectors of time values for xPulses.}

\item{yPulses}{An optional data frame for values to be plotted against xPulses.}

\item{yTime}{An optional time for yPulses.  This must be present if yPulses are used.}

\item{normalization}{A string that indicates the normalization type.  Options: "none", "height", "area".}
}
\value{
A plot of the pulses.
}
\description{
Plots pulses in 2D, 3D waterfall or sequence based on users choice.
}
\details{
This function combines pulses in a data frame that can be used graphically.
Several inputs are based on the user, for example, normalization of the pulses, color ramping and axis labels.
Options on the type of plot include "2D", "3D", "Sequence" and "Heatmap"
}
\examples{
data("pulseData")

xPulses = pulseData$`32AMU`$pulses
xTime = pulseData$`32AMU`$time
plotPulse(xPulses, xTime)






}
\seealso{
\code{\link{pulseNorm}}
}
