% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseAverage.R
\name{pulseAverage}
\alias{pulseAverage}
\title{Pulse Average}
\usage{
pulseAverage(TAPexperiment, gasName, pulseRange = NULL)
}
\arguments{
\item{TAPexperiment}{A set of pre-processing parameters based on the Y-Procedure}

\item{gasName}{Blah}
}
\value{
Blah
}
\description{
Pulse Average
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")

TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = pulseAverage(TAPexperiment, "AMU40", 1:10)

plot(TAPexperiment$AMU40avg$matrix[[1]])

}
