% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseNormalization.R
\name{pulseNorm}
\alias{pulseNorm}
\title{Pulse Normalization}
\usage{
pulseNorm(pulseVector, timeVector, normalization = "none")
}
\arguments{
\item{pulseVector}{A vector of pulse values.}

\item{timeVector}{A vector of time values.}

\item{normalization}{A string that indicates the normalization type.  Options: "none", "height", "area".}
}
\value{
A normalized pulse vector.
}
\description{
This function normalizes the plot based on \code{normalization}.  Typically, this function is used to compare the inert gas with the Standard Diffusion Curve.
Options for normalization include "none", "height", and "area".
}
\details{
For height normalization:
\deqn{\frac{pulse(t)}{max(pulse(t))}}

For area normalization:
\deqn{\frac{pulse(t)}{\int pulse(t)dt}}
}
\examples{
data("pulseData")

pulse = pulseData$`32AMU`$pulses[,33]
timeVector = pulseData$`32AMU`$time
pulseNorm(pulse, timeVector, "area")



}
