% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactionRate.R
\name{reactionRate}
\alias{reactionRate}
\title{Reaction Rate}
\usage{
reactionRate(TAPexperiment, gasName, method = "y")
}
\arguments{
\item{TAPexperiment}{A set of pre-processing parameters based on the Y-Procedure}

\item{gasName}{Blah}

\item{method}{Blah}
}
\value{
Blah
}
\description{
Blah
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")

TAPexperiment$AMU40$options$isProduct = T

TAPexperiment = reactionRate(TAPexperiment, "AMU40")
plot(TAPexperiment$AMU40rate$matrix[[5]])

TAPexperiment = reactionRate(TAPexperiment, "AMU40", method = "g")
plot(TAPexperiment$AMU40rate$matrix[[5]])
}
