% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactivities.R
\name{reactivities}
\alias{reactivities}
\title{Reactivities for TAP}
\usage{
reactivities(TAPexperiment, reactantGasName, productGasName)
}
\arguments{
\item{TAPexperiment}{A set of pre-processing parameters based on the Y-Procedure}

\item{reactantGasName}{Blah}

\item{productGasName}{Blah}
}
\value{
Blah
}
\description{
Reactivities for TAP
}
\details{
blah
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = reactivities(TAPexperiment, "AMU40", "AMU40")
df = momentsToDataFrame(TAPexperiment, "AMU40")
plot(df$R0)

}
