% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTAP.R
\name{readTAP}
\alias{readTAP}
\title{Read Excel TDMS}
\usage{
readTAP(dataPath)
}
\arguments{
\item{dataPath}{This is a string path to an .tdms file.}
}
\value{
A collection of TAP objects that describe the experiment performed in the reactor.
}
\description{
Reads in an Excel .xlsx of the original TDMS file.  This requires that data has been previously converted to .xlsx by the tdms addon in Excel.  This is the function that will be called when an user uploads data.
}
\details{
This function is based on R package rio.
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
names(TAPexperiment)
}
