% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTAPsolver.R
\name{readTAPsolver}
\alias{readTAPsolver}
\title{Read TAPsolver Output}
\usage{
readTAPsolver(dataPath)
}
\arguments{
\item{dataPath}{This is a string path to the output folder of the TAPsolver Experiment}
}
\value{
A collection of TAP objects that describe the experiment performed in the reactor.
}
\description{
Reads the output of the TAPsolver and turns it into the required TAPstructure.
}
\details{
This function is based on R package RJSONIO.
}
\examples{

print("none")

}
