% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustMode.R
\name{robustMode}
\alias{robustMode}
\title{Robust Mode Calculation}
\usage{
robustMode(tempPulse)
}
\arguments{
\item{tempPulse}{a temporary Pulse}
}
\value{
Blah
}
\description{
Roughly based on "On a fast, robust estimator of the mode: Comparisons to other robust estimators with applications"
}
\details{
blah
}
\examples{
data("pulseData")


}
