% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDC.R
\name{SDC}
\alias{SDC}
\title{Standard Diffusion Curve}
\usage{
SDC(TAPexperiment, gasName)
}
\arguments{
\item{diffusion}{Estimated diffusion coefficient.}

\item{tempTime}{A vector of time values.}

\item{L}{The length of the reactor.}

\item{bedPorosity}{Bed Porosity of the reactor.}
}
\value{
A data frame of the pulse and the standard diffusion values based on time.
}
\description{
Create a standard diffusion curve based on the inert gas to determine if the procedure is under the Knudsen diffusion regime.
}
\details{
\deqn{ \frac{1}{2} \pi M0/M1 \sum_{j=0}^{2000} (-1)^j (2j + 1) exp(-0.5 \pi^2 (j + 0.5)^2 M0/M1 )}
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = SDC(TAPexperiment, "AMU40")
plot(TAPexperiment$AMU40SDC$matrix[[5]])
}
