% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDiffusion.R
\name{setDiffusion}
\alias{setDiffusion}
\title{Set Diffusion Coefficient based on moments}
\usage{
setDiffusion(TAPexperiment, gasName)
}
\arguments{
\item{TAPexperiment}{A set of pre-processing parameters based on the Y-Procedure}

\item{gasName}{Blah}
}
\value{
Blah
}
\description{
Blah
}
\details{
blah
}
\examples{
data("pulseData")


}
