% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMoments.R
\name{writeMoments}
\alias{writeMoments}
\title{Write Excel TAP file}
\usage{
writeMoments(TAPexperiment, dataPath)
}
\arguments{
\item{dataPath}{This is a string path to an .xlsx file.}
}
\value{
A collection of TAP objects that describe the experiment performed in the reactor.
}
\description{
Write an Excel .xlsx of the original TDMS file.
}
\details{
This function is based on R package rio.
}
\examples{
TAPexperiment = readTAP("data/pumpProbe.tdms")
TAPexperiment$AMU40$options$baselineStart = 3.9
TAPexperiment$AMU40$options$baselineStart = 4
TAPexperiment$AMU40$options$baselineType = "mean"

TAPexperiment = baselineCorrect(TAPexperiment, "AMU40")
TAPexperiment = moments(TAPexperiment, "AMU40")
writeMoments(TAPexperiment, "data/pumpProbeMoments.xlsx")
}
